+++ src/core/output/generators/files/index.file.ts
@@ -44,7 +44,7 @@ export function createIndexFile(): string {
     ${returnIfTrue(
       i18n,
       `export {useLocalePath, useLocaleRoute} from './__i18n-router';
-      export type {TypedToLocalePath, TypedLocaleRoute, I18nLocales} from './__i18n-router';`
+      export type {TypedToLocalePath, TypedLocaleRoute, I18nLocales, NuxtLocaleRoute} from './__i18n-router';`
     )}