+++ test/e2e/sample-project/[classic].spec.ts
@@ -1,5 +1,0 @@
-import { setupNuxtTestWithConfig } from './utils';
-
-setupNuxtTestWithConfig('classic', {
-  modules: ['nuxt-typed-router'],
-});