+++ test/e2e/sample-project/[withStrict].spec.ts
@@ -1,8 +1,0 @@
-import { setupNuxtTestWithConfig } from './utils';
-
-setupNuxtTestWithConfig('withStrict', {
-  modules: ['nuxt-typed-router'],
-  nuxtTypedRouter: {
-    strict: true,
-  },
-});