+++ test/e2e/sample-project/allConfigs.spec.ts
@@ -1,0 +1,11 @@
+import { setupNuxtTestWithConfig } from './utils';
+
+describe('Testing with different configs', async () => {
+  try {
+    await setupNuxtTestWithConfig('classic');
+    await setupNuxtTestWithConfig('withStrict');
+    await setupNuxtTestWithConfig('withPlugin');
+  } catch (e: any) {
+    throw new Error(e);
+  }
+});