+++ test/fixtures/complex/src/pages/user/index.vue
@@ -1,1 +1,0 @@
-<template><span>H1</span></template>