+++ test/fixtures/simple/pages/user/[id]/posts.vue
@@ -1,1 +1,0 @@
-<template><nuxt-page /></template>