+++ test/samples-config/classic.ts
@@ -1,0 +1,5 @@
+import { NuxtConfig } from '@nuxt/schema';
+
+const config: NuxtConfig = {};
+
+export default config;