+++ test/samples-config/withStrict.ts
@@ -1,0 +1,8 @@
+import { NuxtConfig } from '@nuxt/schema';
+
+const config: NuxtConfig = {
+  nuxtTypedRouter: {
+    strict: true,
+  },
+};
+export default config;