+++ test/samples/[withStrict].config.ts
@@ -1,6 +1,0 @@
-export default {
-  modules: ['nuxt-typed-router'],
-  nuxtTypedRouter: {
-    strict: true,
-  },
-};