+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file. See [standa
 
 ## [4.0.0](https://github.com/victorgarciaesgi/nuxt-typed-router/compare/v3.5.0...v4.0.0) (2023-12-07)
 
+New features
+
+- `ignoreRoutes` option now support Globs pattern
+- Simplification of `NuxtRoute` type, removing the need for generics
+
+Fixs
+
+- Fixed strict:true errors when using `router.push('xxx')`
+- Fixed import of `NuxtLocaleRoute`
+
+
+
+
 Changelog
 All notable changes to this project will be documented in this file.